requirements.txt
flask
edge-tts
pdfplumber
python-docx